requirements.txt
numpy==1.26.4
psutil==5.9.8
tqdm==4.66.4
pyyaml==6.0.1
requests==2.32.3
protobuf==4.25.3

sentence-transformers==2.6.1
transformers==4.41.2
torch==2.2.2

faiss-cpu==1.7.4

datasets==2.19.1
huggingface-hub==0.23.0

pdfplumber==0.11.0
python-docx==1.1.0
python-pptx==0.6.23

pillow==10.3.0
pytesseract==0.3.10
opencv-python==4.9.0.80

langdetect==1.0.9

openai-whisper==20231117
moviepy==1.0.3

onnx==1.16.1
onnxruntime==1.18.0

openvino==2024.2.0
openvino-tokenizers==2024.2.0
openvino-genai==2024.2.0

optimum-intel==1.18.0
nncf==2.11.0